requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
tensorflow-cpu==2.12.0
numpy==1.23.5
opencv-python-headless==4.6.0.66
pillow==9.2.0
python-multipart==0.0.6
pydantic==1.10.7 